requirements.txt
numpy==1.21.6
pandas==1.0.1
PyYAML==6.0.1
matplotlib==3.5.3
